x/cosmonova/types/message_make_request_test.go
package types

import (
	"testing"

	"cosmonova/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgMakeRequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgMakeRequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgMakeRequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgMakeRequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
